requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
databases==0.7.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.101.1
greenlet==2.0.2
httpcore==0.17.3
httpx==0.24.1
h11==0.14.0
idna==3.4
passlib==1.7.4
Pillow==9.5.0
pyasn1==0.5.0
psycopg2-binary==2.9.7
pycparser==2.21
pydantic==2.1.1
pydantic-core==2.4.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlalchemy==1.4.49
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
